requirements/base.txt
Django==2.1.5
python-decouple==3.1
pytz==2018.9
django-crispy-forms==1.7.2
Pillow==5.4.1
requests==2.21.0
django-tables2==2.0.4
beautifulsoup4==4.7.1
lxml==4.3.1
psycopg2-binary
celery>=4.2<4.3
redis==3.2.0
social-auth-app-django==3.1.0
django-extensions==2.1.4
djangorestframework
django-compressor==2.2
